receivable_type model: move to separate file
