+++ leasing/filters.py
@@ -6,12 +6,13 @@ from django_filters.rest_framework import FilterSet, filters
 from rest_framework.filters import OrderingFilter
 
 from leasing.models import CollectionCourtDecision, CollectionLetter, CollectionNote
-from leasing.models.invoice import InvoiceNote, InvoiceRow, InvoiceSet, ReceivableType
+from leasing.models.invoice import InvoiceNote, InvoiceRow, InvoiceSet
 from leasing.models.land_use_agreement import (
     LandUseAgreementInvoice,
     LandUseAgreementInvoiceRow,
     LandUseAgreementInvoiceSet,
 )
+from leasing.models.receivable_type import ReceivableType
 
 from .models import (
     Comment,
